Pre-fill apparence fixed

+++ VultisigApp/VultisigApp/Services/LiFi/LiFiService.swift
@@ -64,7 +64,7 @@ private extension LiFiService {
         struct Estimate: Codable {
             let toAmount: String
             let toAmountMin: String
-            let executionDuration: Int
+            let executionDuration: Decimal
         }
         struct TransactionRequest: Codable {
             let data: String
